refactor: use controller in survey chart, remove stream query
