+++ lib/features/journal/ui/widgets/entry_details/survey_summary.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:lotti/classes/journal_entities.dart';
+import 'package:lotti/features/dashboards/ui/widgets/charts/dashboard_survey_chart.dart';
+import 'package:lotti/features/dashboards/ui/widgets/charts/dashboard_survey_data.dart';
 import 'package:lotti/themes/theme.dart';
-import 'package:lotti/widgets/charts/dashboard_survey_chart.dart';
-import 'package:lotti/widgets/charts/dashboard_survey_data.dart';
 import 'package:lotti/widgets/charts/utils.dart';
 
 class SurveySummary extends StatelessWidget {
