+++ lib/services/db_notification.dart
@@ -38,3 +38,4 @@ class UpdateNotifications {
 const habitCompletionNotification = 'HABIT_COMPLETION';
 const textEntryNotification = 'TEXT_ENTRY';
 const taskNotification = 'TASK';
+const surveyNotification = 'SURVEY';
